+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - `OrientedRectangle.GetVertices()` now returns vertices in the correct order. The 2nd and 3rd vertices were swapped.
+- Fixed `Geometry2D.LineCircle` implementation. It did not detect all intersections correctly.
 
 ## 0.11.2